Rename and reorganize some things
