+++ eventstore/jpa/src/main/java/org/occurrent/eventstore/jpa/EventLogOperations.java
@@ -1,9 +1,16 @@
-package org.occurrent.eventstore.jpa.operations;
+package org.occurrent.eventstore.jpa;
 
+import org.occurrent.eventstore.jpa.mixins.EventLogFilterMixin;
+import org.occurrent.eventstore.jpa.mixins.EventLogSortingMixin;
 import org.springframework.data.jpa.domain.Specification;
 
-public interface EventLogOperations<T>
-    extends EventLogFilterOperations<T>, EventLogSortingOperations<T> {
+/**
+ * Combines all the common operations interfaces into a single type.
+ *
+ * @param <T> the concrete type that the cloud event is stored at. This is a hibernate managed
+ *     entity.
+ */
+public interface EventLogOperations<T> extends EventLogFilterMixin<T>, EventLogSortingMixin<T> {
   default Specification<T> byStreamId(String streamId) {
     return (root, query, criteriaBuilder) -> criteriaBuilder.equal(expressStreamId(root), streamId);
   }
