+++ eventstore/jpa/src/main/java/org/occurrent/eventstore/jpa/JPAEventStore.java
@@ -19,7 +19,6 @@ import org.occurrent.eventstore.api.blocking.EventStore;
 import org.occurrent.eventstore.api.blocking.EventStoreOperations;
 import org.occurrent.eventstore.api.blocking.EventStoreQueries;
 import org.occurrent.eventstore.api.blocking.EventStream;
-import org.occurrent.eventstore.jpa.operations.EventLogOperations;
 import org.occurrent.filter.Filter;
 import org.occurrent.functionalsupport.internal.FunctionalSupport;
 import org.springframework.data.domain.PageRequest;
