+++ eventstore/jpa/src/main/java/org/occurrent/eventstore/jpa/mixins/EventLogFilterMixin.java
@@ -1,10 +1,17 @@
-package org.occurrent.eventstore.jpa.operations;
+package org.occurrent.eventstore.jpa.mixins;
 
 import java.util.function.Supplier;
 import org.occurrent.filter.Filter;
 import org.springframework.data.jpa.domain.Specification;
 
-public interface EventLogFilterOperations<T> extends EventLogConditionOperations<T> {
+/**
+ * Maps instances of {@link Filter} to instances of {@link Specification}. This interface is
+ * completely implemented but consumers can override it if they desire.
+ *
+ * @param <T> the concrete type that the cloud event is stored at. This is a hibernate managed
+ *     entity.
+ */
+public interface EventLogFilterMixin<T> extends EventLogConditionMixin<T> {
   default Specification<T> byFilter(Filter filter) {
     return byFilter(null, filter);
   }
